requirements.txt
flask
requests
json
gunicorn
google-cloud-aiplatform
google-cloud-secret-manager